fix: Fixing api get list categories, adding search

+++ src/main/java/com/example/clinic_management_system/service/CategoryServiceImpl.java
@@ -4,10 +4,12 @@ import com.example.clinic_management_system.dto.request.CategoryRequest;
 import com.example.clinic_management_system.dto.response.CategoryOptionResponse;
 import com.example.clinic_management_system.dto.response.CategoryResponse;
 import com.example.clinic_management_system.entity.Category;
+import com.example.clinic_management_system.entity.Medicine;
 import com.example.clinic_management_system.exception.BadRequestException;
 import com.example.clinic_management_system.exception.ResourceNotFoundException;
 import com.example.clinic_management_system.mapper.CategoryMapper;
 import com.example.clinic_management_system.repository.CategoryRepository;
+import com.example.clinic_management_system.repository.MedicineRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -23,6 +25,7 @@ import java.util.Optional;
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final MedicineRepository medicineRepository;
 
     @Override
     public void createCategory(CategoryRequest categoryRequest) {
@@ -59,17 +62,23 @@ public class CategoryServiceImpl implements CategoryService {
             throw new ResourceNotFoundException("Danh mục không tồn tại");
         }
 
-        // 2. Xóa danh mục
+        // 2. Kiểm tra xem có thuốc nào đang tồn tại với danh mục này không
+        boolean medicineExist = medicineRepository.existsByCategoryId(categoryId);
+        if (medicineExist) {
+            throw new BadRequestException("Không thể xóa, danh mục hiện tại đang có thuốc tồn tại");
+        }
+
+        // 3. Xóa danh mục
         categoryRepository.delete(category.get());
     }
 
     @Override
-    public Page<CategoryResponse> getAllCategories(int page, int size) {
+    public Page<CategoryResponse> getAllCategories(int page, int size, String search) {
         // 1. Tạo đối tượng phân trang
         Pageable pageable = PageRequest.of(page, size, Sort.by("createdAt").descending());
 
         // 2. Query trả về data
-        Page<Category> categories = categoryRepository.findAll(pageable);
+        Page<Category> categories = categoryRepository.findAllCategories(search, pageable);
 
         // 3. Mapping data
         Page<CategoryResponse> categoryResponses = categories.map(categoryMapper::toResponse);
